Delete TS and JS files.

+++ src/models/DefensePlayer.cs
@@ -2,7 +2,7 @@ namespace MaddenImporter
 {
 	public class DefensePlayer : Player
 	{
-		public int Interceptions { get; set; }
+		public float Interceptions { get; set; }
 		public int InterceptionYards { get; set; }
 		public int InterceptionTouchdowns { get; set; }
 		public int LongestInterceptionReturn { get; set; }
@@ -11,7 +11,7 @@ namespace MaddenImporter
 		public int FumblesRecovered { get; set; }
 		public int FumbleYards { get; set; }
 		public int FumbleTouchdowns { get; set; }
-		public int Sacks { get; set; }
+		public float Sacks { get; set; }
 		public int AssistedTackles { get; set; }
 		public int TacklesForLoss { get; set; }
 		public int Safety { get; set; }
